feat(isEmail): add host_whitelist option
